src/test/javascript/spec/app/entities/shoppingcart/cart-product-relation/cart-product-relation.service.spec.ts
/* tslint:disable max-line-length */
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { HttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { take, map } from 'rxjs/operators';
import { CartProductRelationService } from 'app/entities/shoppingcart/cart-product-relation/cart-product-relation.service';
import { ICartProductRelation, CartProductRelation, ProductStatus } from 'app/shared/model/shoppingcart/cart-product-relation.model';

describe('Service Tests', () => {
  describe('CartProductRelation Service', () => {
    let injector: TestBed;
    let service: CartProductRelationService;
    let httpMock: HttpTestingController;
    let elemDefault: ICartProductRelation;
    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule]
      });
      injector = getTestBed();
      service = injector.get(CartProductRelationService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new CartProductRelation(0, ProductStatus.RESOLVED, 0, 0, 0);
    });

    describe('Service methods', async () => {
      it('should find an element', async () => {
        const returnedFromService = Object.assign({}, elemDefault);
        service
          .find(123)
          .pipe(take(1))
          .subscribe(resp => expect(resp).toMatchObject({ body: elemDefault }));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(JSON.stringify(returnedFromService));
      });

      it('should create a CartProductRelation', async () => {
        const returnedFromService = Object.assign(
          {
            id: 0
          },
          elemDefault
        );
        const expected = Object.assign({}, returnedFromService);
        service
          .create(new CartProductRelation(null))
          .pipe(take(1))
          .subscribe(resp => expect(resp).toMatchObject({ body: expected }));
        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(JSON.stringify(returnedFromService));
      });

      it('should update a CartProductRelation', async () => {
        const returnedFromService = Object.assign(
          {
            productStatus: 'BBBBBB',
            productCount: 1,
            couponId: 1,
            categoryId: 1
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);
        service
          .update(expected)
          .pipe(take(1))
          .subscribe(resp => expect(resp).toMatchObject({ body: expected }));
        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(JSON.stringify(returnedFromService));
      });

      it('should return a list of CartProductRelation', async () => {
        const returnedFromService = Object.assign(
          {
            productStatus: 'BBBBBB',
            productCount: 1,
            couponId: 1,
            categoryId: 1
          },
          elemDefault
        );
        const expected = Object.assign({}, returnedFromService);
        service
          .query(expected)
          .pipe(
            take(1),
            map(resp => resp.body)
          )
          .subscribe(body => expect(body).toContainEqual(expected));
        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(JSON.stringify([returnedFromService]));
        httpMock.verify();
      });

      it('should delete a CartProductRelation', async () => {
        const rxPromise = service.delete(123).subscribe(resp => expect(resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
